Refactor: Rename internal_tools to plugins and update all references
