+++ agent/core_agent/plugins/artifacts/plugin.py
