+++ agent/core_agent/plugins/artifacts/schemas.py
